feat(cloud): move reference commands into their product commands
