+++ doc/ovhcloud_cloud_loadbalancer.md
@@ -37,9 +37,11 @@ Manage loadbalancers in the given cloud project
 * [ovhcloud cloud loadbalancer delete](ovhcloud_cloud_loadbalancer_delete.md)	 - Delete a specific loadbalancer
 * [ovhcloud cloud loadbalancer edit](ovhcloud_cloud_loadbalancer_edit.md)	 - Edit the given loadbalancer
 * [ovhcloud cloud loadbalancer get](ovhcloud_cloud_loadbalancer_get.md)	 - Get a specific loadbalancer
+* [ovhcloud cloud loadbalancer get-flavor](ovhcloud_cloud_loadbalancer_get-flavor.md)	 - Get details of a specific loadbalancer flavor
 * [ovhcloud cloud loadbalancer health-monitor](ovhcloud_cloud_loadbalancer_health-monitor.md)	 - Manage health monitors of loadbalancers
 * [ovhcloud cloud loadbalancer l7policy](ovhcloud_cloud_loadbalancer_l7policy.md)	 - Manage L7 policies of loadbalancers
 * [ovhcloud cloud loadbalancer list](ovhcloud_cloud_loadbalancer_list.md)	 - List your loadbalancers
+* [ovhcloud cloud loadbalancer list-flavors](ovhcloud_cloud_loadbalancer_list-flavors.md)	 - List available loadbalancer flavors in the given cloud project
 * [ovhcloud cloud loadbalancer listener](ovhcloud_cloud_loadbalancer_listener.md)	 - Manage listeners of loadbalancers
 * [ovhcloud cloud loadbalancer log](ovhcloud_cloud_loadbalancer_log.md)	 - Manage loadbalancer logs
 * [ovhcloud cloud loadbalancer pool](ovhcloud_cloud_loadbalancer_pool.md)	 - Manage pools of loadbalancers
