+++ doc/ovhcloud_cloud_managed-analytics_list-node-flavors.md
@@ -1,11 +1,22 @@
-## ovhcloud cloud reference managed-database
+## ovhcloud cloud managed-analytics list-node-flavors
 
-Fetch managed database reference data in the given cloud project
+List available analytics node flavors in the given cloud project
+
+```
+ovhcloud cloud managed-analytics list-node-flavors [flags]
+```
 
 ### Options
 
 ```
-  -h, --help   help for managed-database
+      --filter stringArray   Filter results by any property using https://github.com/PaesslerAG/gval syntax
+                             Examples:
+                               --filter 'state=="running"'
+                               --filter 'name=~"^my.*"'
+                               --filter 'nested.property.subproperty>10'
+                               --filter 'startDate>="2023-12-01"'
+                               --filter 'name=~"something" && nbField>10'
+  -h, --help                 help for list-node-flavors
 ```
 
 ### Options inherited from parent commands
@@ -30,8 +41,5 @@ Fetch managed database reference data in the given cloud project
 
 ### SEE ALSO
 
-* [ovhcloud cloud reference](ovhcloud_cloud_reference.md)	 - Fetch reference data in the given cloud project
-* [ovhcloud cloud reference managed-database list-engines](ovhcloud_cloud_reference_managed-database_list-engines.md)	 - List available database engines in the given cloud project
-* [ovhcloud cloud reference managed-database list-node-flavors](ovhcloud_cloud_reference_managed-database_list-node-flavors.md)	 - List available database node flavors in the given cloud project
-* [ovhcloud cloud reference managed-database list-plans](ovhcloud_cloud_reference_managed-database_list-plans.md)	 - List available database plans in the given cloud project
+* [ovhcloud cloud managed-analytics](ovhcloud_cloud_managed-analytics.md)	 - Manage managed analytics services in the given cloud project
