+++ doc/ovhcloud_cloud_managed-database_list-engines.md
@@ -1,11 +1,22 @@
-## ovhcloud cloud reference managed-analytics
+## ovhcloud cloud managed-database list-engines
 
-Fetch managed analytics reference data in the given cloud project
+List available database engines in the given cloud project
+
+```
+ovhcloud cloud managed-database list-engines [flags]
+```
 
 ### Options
 
 ```
-  -h, --help   help for managed-analytics
+      --filter stringArray   Filter results by any property using https://github.com/PaesslerAG/gval syntax
+                             Examples:
+                               --filter 'state=="running"'
+                               --filter 'name=~"^my.*"'
+                               --filter 'nested.property.subproperty>10'
+                               --filter 'startDate>="2023-12-01"'
+                               --filter 'name=~"something" && nbField>10'
+  -h, --help                 help for list-engines
 ```
 
 ### Options inherited from parent commands
@@ -30,8 +41,5 @@ Fetch managed analytics reference data in the given cloud project
 
 ### SEE ALSO
 
-* [ovhcloud cloud reference](ovhcloud_cloud_reference.md)	 - Fetch reference data in the given cloud project
-* [ovhcloud cloud reference managed-analytics list-engines](ovhcloud_cloud_reference_managed-analytics_list-engines.md)	 - List available analytics engines in the given cloud project
-* [ovhcloud cloud reference managed-analytics list-node-flavors](ovhcloud_cloud_reference_managed-analytics_list-node-flavors.md)	 - List available analytics node flavors in the given cloud project
-* [ovhcloud cloud reference managed-analytics list-plans](ovhcloud_cloud_reference_managed-analytics_list-plans.md)	 - List available analytics plans in the given cloud project
+* [ovhcloud cloud managed-database](ovhcloud_cloud_managed-database.md)	 - Manage managed database services in the given cloud project
