+++ doc/ovhcloud_cloud_managed-database_list-node-flavors.md
@@ -1,9 +1,9 @@
-## ovhcloud cloud reference rancher list-versions
+## ovhcloud cloud managed-database list-node-flavors
 
-List available Rancher versions in the given cloud project
+List available database node flavors in the given cloud project
 
 ```
-ovhcloud cloud reference rancher list-versions [flags]
+ovhcloud cloud managed-database list-node-flavors [flags]
 ```
 
 ### Options
@@ -16,8 +16,7 @@ ovhcloud cloud reference rancher list-versions [flags]
                                --filter 'nested.property.subproperty>10'
                                --filter 'startDate>="2023-12-01"'
                                --filter 'name=~"something" && nbField>10'
-  -h, --help                 help for list-versions
-  -r, --rancher-id string    Rancher service ID to filter available versions
+  -h, --help                 help for list-node-flavors
 ```
 
 ### Options inherited from parent commands
@@ -42,5 +41,5 @@ ovhcloud cloud reference rancher list-versions [flags]
 
 ### SEE ALSO
 
-* [ovhcloud cloud reference rancher](ovhcloud_cloud_reference_rancher.md)	 - Fetch Rancher reference data in the given cloud project
+* [ovhcloud cloud managed-database](ovhcloud_cloud_managed-database.md)	 - Manage managed database services in the given cloud project
