+++ doc/ovhcloud_cloud_managed-rancher_list-versions.md
@@ -1,11 +1,23 @@
-## ovhcloud cloud reference loadbalancer
+## ovhcloud cloud managed-rancher list-versions
 
-Fetch loadbalancer reference data in the given cloud project
+List available Rancher versions in the given cloud project
+
+```
+ovhcloud cloud managed-rancher list-versions [flags]
+```
 
 ### Options
 
 ```
-  -h, --help   help for loadbalancer
+      --filter stringArray   Filter results by any property using https://github.com/PaesslerAG/gval syntax
+                             Examples:
+                               --filter 'state=="running"'
+                               --filter 'name=~"^my.*"'
+                               --filter 'nested.property.subproperty>10'
+                               --filter 'startDate>="2023-12-01"'
+                               --filter 'name=~"something" && nbField>10'
+  -h, --help                 help for list-versions
+  -r, --rancher-id string    Rancher service ID to filter available versions
 ```
 
 ### Options inherited from parent commands
@@ -30,7 +42,5 @@ Fetch loadbalancer reference data in the given cloud project
 
 ### SEE ALSO
 
-* [ovhcloud cloud reference](ovhcloud_cloud_reference.md)	 - Fetch reference data in the given cloud project
-* [ovhcloud cloud reference loadbalancer get-flavor](ovhcloud_cloud_reference_loadbalancer_get-flavor.md)	 - Get details of a specific loadbalancer flavor
-* [ovhcloud cloud reference loadbalancer list-flavors](ovhcloud_cloud_reference_loadbalancer_list-flavors.md)	 - List available loadbalancer flavors in the given cloud project
+* [ovhcloud cloud managed-rancher](ovhcloud_cloud_managed-rancher.md)	 - Manage Rancher services in the given cloud project
