--- doc/ovhcloud_cloud_reference.md
@@ -1,41 +1,0 @@
-## ovhcloud cloud reference
-
-Fetch reference data in the given cloud project
-
-### Options
-
-```
-      --cloud-project string   Cloud project ID
-  -h, --help                   help for reference
-```
-
-### Options inherited from parent commands
-
-```
-  -d, --debug            Activate debug mode (will log all HTTP requests details)
-  -e, --ignore-errors    Ignore errors in API calls when it is not fatal to the execution
-  -o, --output string    Output format: json, yaml, interactive, or a custom format expression (using https://github.com/PaesslerAG/gval syntax)
-                         Examples:
-                           --output json
-                           --output yaml
-                           --output interactive
-                           --output 'id' (to extract a single field)
-                           --output 'nested.field.subfield' (to extract a nested field)
-                           --output '[id, "name"]' (to extract multiple fields as an array)
-                           --output '{"newKey": oldKey, "otherKey": nested.field}' (to extract and rename fields in an object)
-                           --output 'name+","+type' (to extract and concatenate fields in a string)
-                           --output '(nbFieldA + nbFieldB) * 10' (to compute values from numeric fields)
-      --profile string   Use a specific profile from the configuration file
-```
-
-### SEE ALSO
-
-* [ovhcloud cloud](ovhcloud_cloud.md)	 - Manage your projects and services in the Public Cloud universe (MKS, MPR, MRS, Object Storage...)
-* [ovhcloud cloud reference list-flavors](ovhcloud_cloud_reference_list-flavors.md)	 - List available flavors in the given cloud project
-* [ovhcloud cloud reference list-images](ovhcloud_cloud_reference_list-images.md)	 - List available images in the given cloud project
-* [ovhcloud cloud reference loadbalancer](ovhcloud_cloud_reference_loadbalancer.md)	 - Fetch loadbalancer reference data in the given cloud project
-* [ovhcloud cloud reference managed-analytics](ovhcloud_cloud_reference_managed-analytics.md)	 - Fetch managed analytics reference data in the given cloud project
-* [ovhcloud cloud reference managed-database](ovhcloud_cloud_reference_managed-database.md)	 - Fetch managed database reference data in the given cloud project
-* [ovhcloud cloud reference managed-rancher](ovhcloud_cloud_reference_managed-rancher.md)	 - Fetch Rancher reference data in the given cloud project
-* [ovhcloud cloud reference managed-registry](ovhcloud_cloud_reference_managed-registry.md)	 - Fetch container registry reference data in the given cloud project
-
