--- doc/ovhcloud_cloud_reference_managed-analytics_list-plans.md
@@ -1,45 +1,0 @@
-## ovhcloud cloud reference managed-analytics list-plans
-
-List available analytics plans in the given cloud project
-
-```
-ovhcloud cloud reference managed-analytics list-plans [flags]
-```
-
-### Options
-
-```
-      --filter stringArray   Filter results by any property using https://github.com/PaesslerAG/gval syntax
-                             Examples:
-                               --filter 'state=="running"'
-                               --filter 'name=~"^my.*"'
-                               --filter 'nested.property.subproperty>10'
-                               --filter 'startDate>="2023-12-01"'
-                               --filter 'name=~"something" && nbField>10'
-  -h, --help                 help for list-plans
-```
-
-### Options inherited from parent commands
-
-```
-      --cloud-project string   Cloud project ID
-  -d, --debug                  Activate debug mode (will log all HTTP requests details)
-  -e, --ignore-errors          Ignore errors in API calls when it is not fatal to the execution
-  -o, --output string          Output format: json, yaml, interactive, or a custom format expression (using https://github.com/PaesslerAG/gval syntax)
-                               Examples:
-                                 --output json
-                                 --output yaml
-                                 --output interactive
-                                 --output 'id' (to extract a single field)
-                                 --output 'nested.field.subfield' (to extract a nested field)
-                                 --output '[id, "name"]' (to extract multiple fields as an array)
-                                 --output '{"newKey": oldKey, "otherKey": nested.field}' (to extract and rename fields in an object)
-                                 --output 'name+","+type' (to extract and concatenate fields in a string)
-                                 --output '(nbFieldA + nbFieldB) * 10' (to compute values from numeric fields)
-      --profile string         Use a specific profile from the configuration file
-```
-
-### SEE ALSO
-
-* [ovhcloud cloud reference managed-analytics](ovhcloud_cloud_reference_managed-analytics.md)	 - Fetch managed analytics reference data in the given cloud project
-
