+++ internal/cmd/cloud_instance.go
@@ -495,5 +495,31 @@ There are three ways to define the installation parameters:
 		Args:  cobra.ExactArgs(1),
 	})
 
+	// Reference data commands (flavors and images available for instances)
+	var flavorRegion string
+	flavorListCmd := withFilterFlag(&cobra.Command{
+		Use:   "list-flavors",
+		Short: "List available flavors in the given cloud project",
+		Run: func(_ *cobra.Command, _ []string) {
+			cloud.GetFlavors(flavorRegion)
+		},
+		Args: cobra.NoArgs,
+	})
+	flavorListCmd.Flags().StringVarP(&flavorRegion, "region", "r", "", "Region to filter flavors (e.g., GRA9, BHS5)")
+	instanceCmd.AddCommand(flavorListCmd)
+
+	var imageRegion, imageOsType string
+	imageListCmd := withFilterFlag(&cobra.Command{
+		Use:   "list-images",
+		Short: "List available images in the given cloud project",
+		Run: func(_ *cobra.Command, _ []string) {
+			cloud.GetImages(imageRegion, imageOsType)
+		},
+		Args: cobra.NoArgs,
+	})
+	imageListCmd.Flags().StringVarP(&imageRegion, "region", "r", "", "Region to filter images (e.g., GRA9, BHS5)")
+	imageListCmd.Flags().StringVarP(&imageOsType, "os-type", "t", "", "OS type to filter images (baremetal-linux, bsd, linux, windows)")
+	instanceCmd.AddCommand(imageListCmd)
+
 	cloudCmd.AddCommand(instanceCmd)
 }
