+++ internal/cmd/cloud_loadbalancer.go
@@ -84,6 +84,21 @@ func initCloudLoadbalancerCommand(cloudCmd *cobra.Command) {
 	// Log sub-commands
 	initLogSubCommands(loadbalancerCmd)
 
+	// Reference data commands
+	loadbalancerCmd.AddCommand(withFilterFlag(&cobra.Command{
+		Use:   "list-flavors <region (GRA9, BHS5, ...)>",
+		Short: "List available loadbalancer flavors in the given cloud project",
+		Run:   cloud.ListLoadbalancerFlavors,
+		Args:  cobra.ExactArgs(1),
+	}))
+
+	loadbalancerCmd.AddCommand(&cobra.Command{
+		Use:   "get-flavor <region> <flavor_id>",
+		Short: "Get details of a specific loadbalancer flavor",
+		Run:   cloud.GetLoadbalancerFlavor,
+		Args:  cobra.ExactArgs(2),
+	})
+
 	cloudCmd.AddCommand(loadbalancerCmd)
 }
